move github classes to contributors.github
