+++ src/main/groovy/org/shipkit/internal/notes/contributors/github/ProjectContributorFetcherFunction.java
@@ -1,4 +1,4 @@
-package org.shipkit.internal.notes.contributors;
+package org.shipkit.internal.notes.contributors.github;
 
 import org.json.simple.JsonObject;
 import org.shipkit.internal.notes.model.ProjectContributor;
