+++ src/test/groovy/org/shipkit/internal/notes/contributors/github/ProjectContributorFetcherFunctionTest.groovy
@@ -1,10 +1,9 @@
-package org.shipkit.internal.notes.contributors
+package org.shipkit.internal.notes.contributors.github
 
 import org.json.simple.JsonObject
 import org.shipkit.internal.notes.model.ProjectContributor
 import org.shipkit.internal.notes.util.GitHubObjectFetcher
 import spock.lang.Specification
-
 
 class ProjectContributorFetcherFunctionTest extends Specification {
